Fix grouping of loops/conditions
Previously, the loops/conditions were wrapped in a group. This normally
makes some amount of sense, since there is a need for there to be
wrapping if the line is too long. But the wrapping should only be done
on the condition / argument to the loop/condition. This allows blocks
that are preceded by a comment to be displayed correctly.